pyproject.toml
[tool.poetry]
name = "target-az-parquet"
version = "0.0.1"
description = "Singer target for AzParquet, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Alexander Shabunevich <alexander.shabunevich@gmail.com>"]
keywords = [
    "ELT",
    "AzParquet",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.8"
singer-sdk = { version="~=0.40.0", extras = ["faker"] }
fs-s3fs = { version = "~=1.1.1", optional = true }
requests = "~=2.32.3"
pyarrow = "^17.0.0"
pyarrowfs-adlgen2 = "^0.2.5"
azure-identity = "^1.17.1"

[tool.poetry.dev-dependencies]
pytest = ">=8"
singer-sdk = { version="~=0.40.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.pytest.ini_options]
addopts = '--durations=10'

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true

[tool.ruff]
src = ["target_az_parquet"]
target-version = "py38"

[tool.ruff.lint]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    "COM812",  # missing-trailing-comma
    "ISC001",  # single-line-implicit-string-concatenation
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.isort]
known-first-party = ["target_az_parquet"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core==1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-az-parquet = 'target_az_parquet.target:TargetAzParquet.cli'
